Refactored wireframe ai service
